+++ src/Gateway.php
@@ -30,4 +30,9 @@ class Gateway extends AbstractGateway
     {
         return $this->createRequest('\Omnipay\Zero\Message\Request', $parameters);
     }
+
+    public function purchase(array $parameters = array())
+    {
+        return $this->createRequest('\Omnipay\Zero\Message\Request', $parameters);
+    }
 }