+++ CHANGELOG.md
@@ -5,6 +5,8 @@ Updates should follow the [Keep a CHANGELOG](http://keepachangelog.com/) princip
 ## [Unreleased]
 ### Added
 
+ - Added support for purchase() method
+
 ## 2.0.0
 ### Added